Rename WIP functions to wip_*

+++ wip_plotPoseList.m
@@ -8,9 +8,8 @@ ip = inputParser;
 %%% Define validation methods
 % Pose list must be validated to have 13 columns
 valFcn_PoseList = @(x) ismatrix(x) && size(x, 2) == 13;
-% Figure handles may be provided but only if they are a valid figure handle
-% (duh)
-valFcn_FigureHandle = @(x) ishandle(x) && strcmp(get(x, 'type'), 'figure');
+% Option to 'axes' must be a handle and also a 'axes' handle
+valFcn_Axes = @(x) ishandle(x) && strcmp(get(x, 'type'), 'axes');
 % Allow the plot style to be chosen
 valFcn_PlotStyle = @(x) ischar(x) && any(strcmpi(x, {'Single2D', 'Single2D+3D', 'Multi2D', 'Multi2D+3D', '3D', '3D+Single2D', '3D+Multi2d'}));
 % Integer to determine the column of time can be passed, too
@@ -20,6 +19,8 @@ valFcn_ColumnTime = @(x) isinteger(x) && x >= 0 && x < size(PoseList, 2);
 valFcn_ColumnsXYZ = @(x) isrow(x) && issize(x, 1, 3);
 % Validate the 3d viewport argument is correct
 valFcn_Viewport3D = @(x) isrow(x) && isequal(size(x, 2), 2);
+% Grid may be true, false, 1, 0, 'on', 'off', or 'minor'
+valFcn_Grid = @(x) islogical(x) || ( isequal(x, 1) || isequal(x, 0) ) || any(strcmpi(x, {'on', 'off', 'minor'}));
 
 %%% This fills in the parameters for the function
 % We need the pose list
@@ -28,9 +29,9 @@ addOptional(ip, 'PlotStyle', '3D', valFcn_PlotStyle);
 % % We allow the user to explicitley flag which algorithm to use
 % addOptional(ip, 'Plot3d', false, @islogical);
 % Allow options to be passed to the figure
-% addOptional(ip, 'FigureHandle', false, valFcn_FigureHandle);
+addOptional(ip, 'Axes', gca, valFcn_Axes);
 % Allow the figure to have user-defined properties
-addOptional(ip, 'FigureProperties', {}, @iscell);
+% addOptional(ip, 'FigureProperties', {}, @iscell);
 % Allow the lines drawn to have user-defined properties
 addOptional(ip, 'LineProperties', {}, @iscell);
 % If time column should not be detected automatically, it may be set
@@ -40,9 +41,11 @@ addOptional(ip, 'ColumnTime', 0, valFcn_ColumnTime)
 % assuming them to be in columns [2, 3, 4]
 addOptional(ip, 'ColumnsXYZ', [2, 3, 4], valFcn_ColumnsXYZ);
 % User might want to change the 3D view port
-addOptional(ip, 'Viewport3D', [-20, 6], valFcn_Viewport3D);
+addOptional(ip, 'Viewport3D', [-13, 10], valFcn_Viewport3D);
 % Maybe the direction of the 3D-plot's trajectory is desired, too?
 addOptional(ip, 'PlotDirection3D', false, @islogical);
+% Allow user to choose grid style (either false 'on', 'off', or 'minor'
+addOptional(ip, 'Grid', false, valFcn_Grid);
 
 % Configuratio nfor the input parser
 ip.KeepUnmatched = true;
@@ -58,18 +61,9 @@ parse(ip, PoseList, varargin{:});
 % again
 mPoseList = ip.Results.PoseList;
 cPlotStyle = lower(ip.Results.PlotStyle);
-% bPlot3d = ip.Results.Plot3d;
-clFigureProperties = ip.Results.FigureProperties;
+hAxes = ip.Results.Axes;
+% clFigureProperties = ip.Results.FigureProperties;
 clLineProperties = ip.Results.LineProperties;
-% Check the figure handle provided (if a logical value
-% if ishandle(ip.Results.FigureHandle)
-%     hFig = ip.Results.FigureHandle;
-% % if islogical(ip.Results.FigureHandle) && ~ip.Results.FigureHandle
-%     hFig = figure;
-% else
-%     hFig = ip.Results.FigureHandle;
-% end
-hFig = figure;
 % Set the column that keeps the time here
 iColumnTime = ip.Results.ColumnTime;
 % If column is set to 0, then we need to detect it automatically
@@ -79,6 +73,12 @@ end
 vColumnsXYZ = ip.Results.ColumnsXYZ;
 v3DViewport = ip.Results.Viewport3D;
 b3DPlotDireciton = ip.Results.PlotDirection3D;
+chGrid = ip.Results.Grid;
+
+% If this is a single plot i.e., the given axes does not have any children, then
+% we are completely free at plotting stuff like labels, etc., Otherwise, we will
+% really just plot the robot frame
+bNewPlot = isempty(get(hAxes, 'Children'));
 
 
 
@@ -86,39 +86,46 @@ b3DPlotDireciton = ip.Results.PlotDirection3D;
 
 
 
+
 %% Actual plotting
 % Select the provided figure handle to be the active handle
-figure(hFig);
+axes(hAxes);
+
+% Ensure we are not overriding any previously plotted data in the given axex
+hold(hAxes, 'on');
 
 %%% Plot the pose list, all commands furthermore passed to PLOTPOSELIST
 %%% will be arguments to either plot3 or plot
 switch cPlotStyle
     case 'single2d'
-        plot(NaN, NaN);
-        hold on;
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
-        xlabel('Time $t \left[ \rm s \right]$');
-        ylabel('Position $\left[ \rm m \right]$');
-        legend('$x$', '$y$', '$z$');
+        
+        if bNewPlot
+            ylim(hAxes, ylim().*1.05);
+            xlabel('Time $t \left[ \rm s \right]$');
+            ylabel('Position $\left[ \rm m \right]$');
+            legend('$x$', '$y$', '$z$');
+        end
     case 'single2d+3d'
         % Plot 3d
         subplot(2, 1, 2);
         plot3(mPoseList(:, vColumnsXYZ(1)), mPoseList(:, vColumnsXYZ(2)), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
+        
         xlabel('$x \left[ m \right]$');
         ylabel('$y \left[ m \right]$');
         zlabel('$z \left[ m \right]$');
         
-        grid on;
+        if chGrid
+            grid(hAxes, chGrid);
+        end
+        
         view(v3DViewport);
         
         % Plot [x, y, z] vs t
         subplot(2, 1, 1);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $\left[ \rm m \right]$');
         legend('$x$', '$y$', '$z$');
@@ -126,66 +133,69 @@ switch cPlotStyle
         % Plot x vs t
         subplot(3, 1, 1);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(1)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $x \left[ \rm m \right]$');
         
         % Plot y vs t
         subplot(3, 1, 2);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(2)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $y \left[ \rm m \right]$');
         
         % Plot z vs t
         subplot(3, 1, 3);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $z \left[ \rm m \right]$');
     case 'multi2d+3d'
         % Plot 3d
         subplot(2, 3, [4, 5, 6]);
         plot3(mPoseList(:, vColumnsXYZ(1)), mPoseList(:, vColumnsXYZ(2)), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
+        
         xlabel('$x \left[ m \right]$');
         ylabel('$y \left[ m \right]$');
         zlabel('$z \left[ m \right]$');
         
-        grid on;
+        if chGrid
+            grid(hAxes, chGrid);
+        end
+        
         view(v3DViewport);
         
         % Plot x vs t
         subplot(2, 3, 1);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(1)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $x \left[ \rm m \right]$');
         
         % Plot y vs t
         subplot(2, 3, 2);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(2)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $y \left[ \rm m \right]$');
         
         % Plot z vs t
         subplot(2, 3, 3);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $z \left[ \rm m \right]$');
     case '3d'
         % Plot 3d
         plot3(mPoseList(:, vColumnsXYZ(1)), mPoseList(:, vColumnsXYZ(2)), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
+        
         if b3DPlotDireciton
-            hold on
 %             vIndexArrowStart = 1:(size(mPoseList(:, iColumnTime)) - 1);
 %             vIndexArrowEnd = vIndexArrowStart + 1;
 %             vArrowX = 0.5.*(mPoseList(vIndexArrowEnd, vColumnsXYZ(1)) - mPoseList(vIndexArrowStart, vColumnsXYZ(1)));
@@ -194,31 +204,41 @@ switch cPlotStyle
             quiver3(mPoseList(:, vColumnsXYZ(1)), mPoseList(:, vColumnsXYZ(2)), mPoseList(:, vColumnsXYZ(3)), gradient(mPoseList(:, vColumnsXYZ(1))), gradient(mPoseList(:, vColumnsXYZ(2))), gradient(mPoseList(:, vColumnsXYZ(3))));
 %             quiver3(mPoseList(vIndexArrowStart, vColumnsXYZ(1)), mPoseList(vIndexArrowStart, vColumnsXYZ(2)), mPoseList(vIndexArrowStart, vColumnsXYZ(3)), vArrowX, vArrowY, vArrowZ, 0);
         end
-        axis('tight')
+        
+        xlim(hAxes, xlim().*1.05);
+        ylim(hAxes, ylim().*1.05);
+        zlim(hAxes, zlim().*1.05);
+        
         xlabel('$x \left[ m \right]$');
         ylabel('$y \left[ m \right]$');
         zlabel('$z \left[ m \right]$');
         
-        grid on;
+        if chGrid
+            grid(hAxes, chGrid);
+        end
+        
         view(v3DViewport);
     case '3d+single2d'
         % Plot 3d
         subplot(1, 2, 1);
         plot3(mPoseList(:, vColumnsXYZ(1)), mPoseList(:, vColumnsXYZ(2)), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('$x \left[ m \right]$');
         ylabel('$y \left[ m \right]$');
         zlabel('$z \left[ m \right]$');
         
-        grid on;
+        if chGrid
+            grid(hAxes, chGrid);
+        end
+        
         view(v3DViewport);
         
         % Plot [x, y, z] vs t
         subplot(1, 2, 2);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.05);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $\left[ \rm m \right]$');
         legend('$x$', '$y$', '$z$');
@@ -226,70 +246,54 @@ switch cPlotStyle
         % Plot 3d
         subplot(3, 2, [1, 3, 5]);
         plot3(mPoseList(:, vColumnsXYZ(1)), mPoseList(:, vColumnsXYZ(2)), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
+        
         xlabel('$x \left[ m \right]$');
         ylabel('$y \left[ m \right]$');
         zlabel('$z \left[ m \right]$');
         
-        grid on;
+        if chGrid
+            grid(hAxes, chGrid);
+        end
+        
         view(v3DViewport);
         
         % Plot x vs t
         subplot(3, 2, 2);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(1)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.10);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $x \left[ \rm m \right]$');
         
         % Plot y vs t
         subplot(3, 2, 4);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(2)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.10);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $y \left[ \rm m \right]$');
         
         % Plot z vs t
         subplot(3, 2, 6);
         plot(mPoseList(:, iColumnTime), mPoseList(:, vColumnsXYZ(3)), clLineProperties{:});
-        axis('tight')
-        ylim(ylim().*1.10);
+        
+        ylim(hAxes, ylim().*1.10);
         xlabel('Time $t \left[ \rm s \right]$');
         ylabel('Position $z \left[ \rm m \right]$');
         
     otherwise
 end
-% % Plot in 3D?
-% if bPlot3d
-% % Plot as 2D plot
-% else
-%     plot(mPoseList(:, iColumnTime), mPoseList(:, ColumnsXYZ), clLineProperties{:});
-%     xlabel('Time $t \left[s \right]$');
-%     ylabel('Position $\left[ m \right]$');
-% end
-
-% Put the figure in the foreground
-figure(hFig);
 
 % And draw the image
 drawnow;
-
-if ~isempty(clFigureProperties)
-    % Set the provided figure properties (this *MUST* be done down here because
-    % plot or plo3 will change propertie slike 'Visible', and alike
-    set(hFig, clFigureProperties{:});
-end
 
 
 
 %% Define return values
 
 % First return value will be the handle to the figure
-if nargout > 0
-    if nargout >= 1
-        varargout{1} = hFig;
-    end
+if nargout >= 1
+    varargout{1} = hAxes;
 end
 
 end
